requirements.txt
streamlit
opencv-python-headless
moviepy
torch
transformers
Pillow
psutil
sentence-transformers